Moved all config related functionality into a new ConfigContext.
