+++ src/Drupal/DrupalExtension/Context/DrupalContext.php
@@ -410,50 +410,6 @@ class DrupalContext extends RawDrupalContext implements TranslatableContext {
   }
 
   /**
-   * Sets basic configuration item.
-   *
-   * @param string $name
-   *   The name of the configuration object.
-   * @param string $key
-   *   Identifier to store value in configuration.
-   * @param mixed $value
-   *   Value to associate with identifier.
-   *
-   * @Given I set the configuration item :name with identifier :key to :value
-   */
-  public function setBasicConfig($name, $key, $value) {
-    $this->setConfig($name, $key, $value);
-  }
-
-  /**
-   * Sets complex configuration.
-   *
-   * @param string $name
-   *   The name of the configuration object.
-   * @param string $key
-   *   Identifier to store value in configuration.
-   * @param TableNode $config_table
-   *   The table listing configuration keys and values.
-   *
-   * @Given I set the configuration item :name with identifier :key with values:
-   *
-   * Provide configuration data in the following format:
-   *  | key   | value  |
-   *  | foo   | bar    |
-   */
-  public function setComplexConfig($name, $key, TableNode $config_table) {
-    $value = array();
-    foreach ($config_table->getHash() as $row) {
-      // Allow json values for extra complexity.
-      if (json_decode($row['value'])) {
-        $row['value'] = json_decode($row['value'], TRUE);
-      }
-      $value[$row['key']] = $row['value'];
-    }
-    $this->setConfig($name, $key, $value);
-  }
-
-  /**
    * Pauses the scenario until the user presses a key. Useful when debugging a scenario.
    *
    * @Then (I )break
